Move test datatypes to test library
