+++ memory/ref_counted_test.cc
@@ -1,107 +1,19 @@
 #include "common/memory/ref_counted.h"
 
 #include <cstdint>
-#include <iostream>
 #include <utility>
 
 #include "gtest/gtest.h"
 
+#include "common/test_structures/base_types.h"
+
 namespace common {
 
-namespace {
-
-struct Base {
-  virtual ~Base() = default;
-};
-
-struct NonCopyMovable : public Base {
-  NonCopyMovable(std::size_t value, std::size_t* destructor_count = nullptr)
-      : value_(value), destructor_count_(destructor_count) {
-    if (destructor_count_ != nullptr) {
-      *destructor_count_ = 0;
-    }
-  }
-
-  ~NonCopyMovable() override {
-    if (destructor_count_ != nullptr) {
-      ++(*destructor_count_);
-    }
-  }
-
-  NonCopyMovable(const NonCopyMovable&) = delete;
-  NonCopyMovable& operator=(const NonCopyMovable&) = delete;
-
-  std::size_t value_;
-  std::size_t* const destructor_count_;
-};
-
-struct Movable : public Base {
-  Movable(std::size_t value, std::size_t* destructor_count = nullptr)
-      : value_(value), destructor_count_(destructor_count) {
-    if (destructor_count_ != nullptr) {
-      *destructor_count_ = 0;
-    }
-  }
-
-  ~Movable() override {
-    if (destructor_count_ != nullptr) {
-      ++(*destructor_count_);
-    }
-  }
-
-  Movable(const Movable&) = delete;
-  Movable(Movable&&) = default;
-
-  Movable& operator=(const Movable&) = delete;
-  Movable& operator=(Movable&&) = default;
-
-  std::size_t value_;
-  std::size_t* const destructor_count_;
-};
-
-struct Copyable : public Base {
-  Copyable(std::size_t value, std::size_t* destructor_count = nullptr)
-      : value_(value), destructor_count_(destructor_count) {
-    if (destructor_count_ != nullptr) {
-      *destructor_count_ = 0;
-    }
-  }
-
-  ~Copyable() override {
-    if (destructor_count_ != nullptr) {
-      ++(*destructor_count_);
-    }
-  }
-
-  Copyable(const Copyable&) = default;
-  Copyable(Copyable&&) = delete;
-
-  Copyable& operator=(const Copyable&) = default;
-  Copyable& operator=(Copyable&&) = delete;
-
-  std::size_t value_;
-  std::size_t* const destructor_count_;
-};
-
-struct CopyMovable : public Base {
-  CopyMovable(std::size_t value, std::size_t* destructor_count = nullptr)
-      : value_(value), destructor_count_(destructor_count) {
-    if (destructor_count_ != nullptr) {
-      *destructor_count_ = 0;
-    }
-  }
-
-  ~CopyMovable() override {
-    if (destructor_count_ != nullptr) {
-      ++(*destructor_count_);
-    }
-  }
-
-  std::size_t value_;
-  std::size_t* const destructor_count_;
-};
-
-}  // namespace
+using test_structures::Base;
+using test_structures::NonCopyMovable;
+using test_structures::Movable;
+using test_structures::Copyable;
+using test_structures::CopyMovable;
 
 template <typename T>
 struct RefCountedTest : public testing::Test {
