+++ test_structures/BUILD
@@ -1,0 +1,14 @@
+package(
+    default_visibility = ["//visibility:private"],
+    default_testonly = True,
+)
+
+licenses(["notice"])
+
+cc_library(
+    name = "base_types",
+    hdrs = [
+        "base_types.h",
+    ],
+    visibility = ["//visibility:public"],
+)
